Assessment_Q4
This query estimates the Customer Lifetime Value (CLV) for each user based on their savings activity. CLV helps to measure how valuable a customer is to the business over time. It pulls user ID and full name from the users table, calculates the tenure (how long with the business) for each person in months, counts the number of transactions then estimates the CLV using a formular.

It finally groups and sorts results by estimated CLV in descending order. This query helps in identifying high-value customers for retention strategies, VIP programs, or upselling opportunities.